Remove no variance match simulator

+++ src/main/java/com/roumada/swiftscore/model/FootballClub.java
@@ -13,7 +13,7 @@ public class FootballClub {
     @Id
     private Long id;
     private String name;
-    private float victoryChance;
+    private double victoryChance;
 
     public static class Builder {
         public FootballClub build() {
